+++ ui/src/components/PageScreen.svelte
@@ -301,7 +301,7 @@
                     class="btn btn-circle {$mode == modes.preview
                         ? 'btn-primary btn-transparent'
                         : 'btn-hint btn-transparent txt-hint'}"
-                    use:tooltip={{ position: "top", text: "Preview mode\n(P)" }}
+                    use:tooltip={{ position: "top", text: "Preview mode", sub: "P" }}
                     on:click|preventDefault={() => changeMode(modes.preview)}
                 >
                     <i class="iconoir-play" />
@@ -311,7 +311,7 @@
                         ? 'btn-success btn-transparent'
                         : 'btn-hint btn-transparent txt-hint'}"
                     class:btn-loading={$isLoadingTemplates || $isLoadingHotspots}
-                    use:tooltip={{ position: "top", text: "Hotspots mode\n(H)" }}
+                    use:tooltip={{ position: "top", text: "Hotspots mode", sub: "H" }}
                     on:click|preventDefault={() => changeMode(modes.hotspots)}
                 >
                     <i class="iconoir-square-3d-corner-to-corner" />
@@ -321,7 +321,7 @@
                         ? 'btn-danger btn-transparent'
                         : 'btn-hint btn-transparent txt-hint'}"
                     class:btn-loading={$isLoadingComments}
-                    use:tooltip={{ position: "top", text: "Comments mode\n(C)" }}
+                    use:tooltip={{ position: "top", text: "Comments mode", sub: "C" }}
                     on:click|preventDefault={() => changeMode(modes.comments)}
                 >
                     <i class="iconoir-message-text" />